# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authentication API",
        "description": "API documentation about the authentication endpoints.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user with the provided name, email, and password.",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the user",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "The email address of the user",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "The password of the user",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Bad request. Indicates that the request body is invalid or incomplete."
                    },
                    "422": {
                        "description": "Unprocessable entity. Indicates that the request body is invalid or incomplete."
                    },
                    "500": {
                        "description": "Internal server error. Indicates a server-side problem."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user",
                "description": "Authenticate a user with the provided email and password.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "The email address of the user",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "The password of the user",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully"
                    },
                    "401": {
                        "description": "Unauthorized. Indicates invalid credentials or user not found."
                    }
                }
            }
        },
        "/api/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get all blogs",
                "description": "Returns a list of all blogs.",
                "operationId": "393feb67e712ffe8052c83b54689ceb0",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of all blogs"
                    },
                    "401": {
                        "description": "Unauthorized. Indicates that the user is not authenticated."
                    },
                    "500": {
                        "description": "Internal server error. Indicates a server-side problem."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blog/create": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Create a new blog",
                "description": "Creates a new blog with the provided title, image, and body.",
                "operationId": "4ca9d6012ba5ccb3d6dfb0d95bcb0249",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Blog information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image",
                                    "body"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "The title of the blog",
                                        "type": "string",
                                        "example": "My first blog"
                                    },
                                    "image": {
                                        "description": "The image of the blog",
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "body": {
                                        "description": "The body of the blog",
                                        "type": "string",
                                        "example": "This is the body of my first blog"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog created successfully"
                    },
                    "400": {
                        "description": "Bad request. Indicates that the request body is invalid or incomplete."
                    },
                    "401": {
                        "description": "Unauthorized. Indicates that the user is not authenticated."
                    },
                    "422": {
                        "description": "Unprocessable entity. Indicates that the request body is invalid or incomplete."
                    },
                    "500": {
                        "description": "Internal server error. Indicates a server-side problem."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blog/update/{id}": {
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update a blog by ID",
                "description": "Updates a blog with the provided ID.",
                "operationId": "817dcfded8d33b76b3ac5797b52c7981",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Blog information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image",
                                    "body"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "The title of the blog",
                                        "type": "string",
                                        "example": "My first blog"
                                    },
                                    "image": {
                                        "description": "The image of the blog",
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "body": {
                                        "description": "The body of the blog",
                                        "type": "string",
                                        "example": "This is the body of my first blog"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog updated successfully"
                    },
                    "400": {
                        "description": "Bad request. Indicates that the request body is invalid or incomplete."
                    },
                    "401": {
                        "description": "Unauthorized. Indicates that the user is not authenticated."
                    },
                    "404": {
                        "description": "Not found. Indicates that the blog with the provided ID was not found."
                    },
                    "422": {
                        "description": "Unprocessable entity. Indicates that the request body is invalid or incomplete."
                    },
                    "500": {
                        "description": "Internal server error. Indicates a server-side problem."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blog/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get a blog by ID",
                "description": "Returns a blog with the provided ID.",
                "operationId": "92fdf5c0182582715d2439ee4f758be6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A blog with the provided ID"
                    },
                    "404": {
                        "description": "Not found. Indicates that the blog with the provided ID was not found."
                    },
                    "401": {
                        "description": "Unauthorized. Indicates that the user is not authenticated."
                    },
                    "500": {
                        "description": "Internal server error. Indicates a server-side problem."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}